Move Address.ts to top-level
