+++ src/index.ts
@@ -1,7 +1,8 @@
-export * from "./address/Address";
+export * from "./Address";
 export * from "./agent/Agent";
 export * from "./expression/Expression";
 export * from "./expression/ExpressionRef";
+export * from "./language/Icon";
 export * from "./language/Language";
 export * from "./language/LanguageRef";
 export * from "./language/LanguageContext";
